private/requirements.txt
# Flask Core
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1

# Database & Migrations
Flask-Migrate==4.1.0
SQLAlchemy==2.0.45
SQLAlchemy-Utils==0.42.0
PyMySQL==1.1.2
alembic==1.17.2

# LLM & Vector DB (from original requirements)
openai==2.11.0
chromadb==1.3.6

# Utilities
python-dotenv==1.2.1
requests==2.32.5
email-validator==2.3.0
Pillow==12.0.0
